Refactor calcNodesPageRank
