+++ lib/scala/kg/src/main/scala/data/kg/KgData.scala
@@ -2,7 +2,7 @@ package data.kg
 
 import formats.kg.kgtk.KgtkEdgeWithNodes
 import models.kg.{KgEdge, KgNode, KgPath, KgSource}
-import util.CalcNodesPageRank
+import util.NodePageRankCalculator
 
 import scala.collection.mutable.HashMap
 
@@ -14,7 +14,7 @@ abstract class KgData(edgesUnsorted: List[KgEdge], nodesUnsorted: List[KgNode], 
   val paths = validatePaths(edges, nodesByIdUnranked, pathsUnsorted)
   val sourcesById = (nodesByIdUnranked.flatMap(_._2.sources) ++ edges.flatMap(_.sources)).map(KgSource(_)).map(source => (source.id, source)).toMap
   val sources = sourcesById.values.toList
-  val nodes = CalcNodesPageRank(nodesByIdUnranked.values.toList, edges)
+  val nodes = NodePageRankCalculator(nodesByIdUnranked.values.toList, edges)
   val nodesById = nodes.map{node => (node.id, node)}.toMap
 
   def this(resources: CskgCsvDataResources) =
